// Client/companeros-de-viaje-app/src/app/core/interfaces/groupList.interface.ts
// Generated by https://quicktype.io

export interface GroupList {
  content: Content[];
  pageable: Pageable;
  totalPages: number;
  totalElements: number;
  last: boolean;
  size: number;
  number: number;
  sort: Sort;
  numberOfElements: number;
  first: boolean;
  empty: boolean;
}

export interface Content {
  id: number;
  destination: string;
  departureDate: string;
  returnDate: string;
  itinerary: string;
  budget: number;
  owner: Owner;
  interests: string[];
  minimumNumberOfMembers: number;
  travelers: Owner[];
  totalTravelers: number;
}

export interface Owner {
  username: string;
  name: string;
  profilePicture: null;
}

export interface Pageable {
  pageNumber: number;
  pageSize: number;
  sort: Sort;
  offset: number;
  paged: boolean;
  unpaged: boolean;
}

export interface Sort {
  empty: boolean;
  sorted: boolean;
  unsorted: boolean;
}